# swagger.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "University Portal API",
    "version": "1.0.0",
    "description": "Contract-first shapes for key DTOs and select endpoints. Auto-generated paths will be merged with these definitions."
  },
  "servers": [
    {
      "url": "https://api.university.edu",
      "description": "Production server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/courses/{id}": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Auto-generated for GET /courses/:id",
        "operationId": "get-courses-id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "course.routes.ts"
      },
      "put": {
        "tags": [
          "Courses"
        ],
        "summary": "Auto-generated for PUT /courses/:id",
        "operationId": "put-courses-id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "course.routes.ts"
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "summary": "Auto-generated for DELETE /courses/:id",
        "operationId": "delete-courses-id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "course.routes.ts"
      }
    },
    "/courses/{id}/enroll": {
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Auto-generated for POST /courses/:id/enroll",
        "operationId": "post-courses-id-enroll",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "course.routes.ts"
      }
    },
    "/courses/{id}/unenroll": {
      "delete": {
        "tags": [
          "Courses"
        ],
        "summary": "Auto-generated for DELETE /courses/:id/unenroll",
        "operationId": "delete-courses-id-unenroll",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "course.routes.ts"
      }
    },
    "/courses/{id}/students": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Auto-generated for GET /courses/:id/students",
        "operationId": "get-courses-id-students",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "course.routes.ts"
      }
    },
    "/courses/{id}/materials": {
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Auto-generated for POST /courses/:id/materials",
        "operationId": "post-courses-id-materials",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "course.routes.ts"
      },
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Auto-generated for GET /courses/:id/materials",
        "operationId": "get-courses-id-materials",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "course.routes.ts"
      }
    },
    "/courses/{id}/materials/{materialId}/download": {
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Auto-generated for POST /courses/:id/materials/:materialId/download",
        "operationId": "post-courses-id-materials-materialId-download",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "materialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "course.routes.ts"
      }
    },
    "/courses/{id}/materials/{materialId}": {
      "delete": {
        "tags": [
          "Courses"
        ],
        "summary": "Auto-generated for DELETE /courses/:id/materials/:materialId",
        "operationId": "delete-courses-id-materials-materialId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "materialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "course.routes.ts"
      }
    },
...
    "/students/courses/{id}": {
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "Auto-generated for GET /students/courses/:id",
        "operationId": "get-students-courses-id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "student.routes.ts"
      }
    },
    "/students/courses/{id}/materials": {
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "Auto-generated for GET /students/courses/:id/materials",
        "operationId": "get-students-courses-id-materials",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "student.routes.ts"
      }
    },
    "/students/courses/{id}/materials/{materialId}/download": {
      "post": {
        "tags": [
          "Students"
        ],
        "summary": "Auto-generated for POST /students/courses/:id/materials/:materialId/download",
        "operationId": "post-students-courses-id-materials-materialId-download",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "materialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "student.routes.ts"
      }
    },
...
    "/lecturers/courses/{id}": {
      "get": {
        "tags": [
          "Lecturers"
        ],
        "summary": "Auto-generated for GET /lecturers/courses/:id",
        "operationId": "get-lecturers-courses-id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "lecturer.routes.ts"
      }
    },
    "/lecturers/courses/{id}/students": {
      "get": {
        "tags": [
          "Lecturers"
        ],
        "summary": "Auto-generated for GET /lecturers/courses/:id/students",
        "operationId": "get-lecturers-courses-id-students",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "lecturer.routes.ts"
      }
    },
    "/lecturers/courses/{id}/materials": {
      "post": {
        "tags": [
          "Lecturers"
        ],
        "summary": "Auto-generated for POST /lecturers/courses/:id/materials",
        "operationId": "post-lecturers-courses-id-materials",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "lecturer.routes.ts"
      }
    },
    "/lecturers/courses/{id}/materials/{materialId}": {
      "delete": {
        "tags": [
          "Lecturers"
        ],
        "summary": "Auto-generated for DELETE /lecturers/courses/:id/materials/:materialId",
        "operationId": "delete-lecturers-courses-id-materials-materialId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "materialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "lecturer.routes.ts"
      }
    },
...
    "/admin/courses/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Auto-generated for GET /admin/courses/:id",
        "operationId": "get-admin-courses-id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "admin.routes.ts"
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Auto-generated for PUT /admin/courses/:id",
        "operationId": "put-admin-courses-id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "admin.routes.ts"
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Auto-generated for DELETE /admin/courses/:id",
        "operationId": "delete-admin-courses-id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-source": "admin.routes.ts"
      }
    },
...
  }
}